moved matrix calculations to Mat4
